python_ocr/requirements.txt
# Python OCR Engine Dependencies
# Install with: pip install -r requirements.txt

opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
